requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
feedparser==6.0.11
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
sgmllib3k==1.0.0
six==1.17.0
typing_extensions==4.12.2
websockets==15.0
yarl==1.18.3
